src/requirements.txt
blinker (1.4)
colorama (0.3.9)
docutils (0.14)
entrypoints (0.2.3)
feedgenerator (1.9)
Jinja2 (2.10)
keyring (12.0.0)
Markdown (2.6.11)
MarkupSafe (1.0)
pelican (3.7.1)
pyftpsync (2.0.0)
Pygments (2.2.0)
python-dateutil (2.7.1)
pytz (2018.3)
six (1.11.0)
smartypants (2.0.1)
typogrify (2.0.7)
Unidecode (1.0.22)
